chore(phai): split agent-run surface into a tiered api/ facade
Replace the single index.ts barrel with five domain-scoped re-export modules under
products/posthog_ai/frontend/api/ (run / primitives / logics / types / tools), split along
dependency and side-effect boundaries so each consumer pulls only its subtree. The headless
lane (api/logics + api/types) stays free of React and the registry; the side-effectful tool
registry is isolated in api/tools. The root index.ts becomes a thin back-compat aggregate.

Add registerToolRenderers(entries) to the registry as the generic per-product tool-card seam,
and make Max its first consumer (registerMaxToolRenderers) rather than a special case.

Migrate every known consumer (tasks, signals inbox, Max, ai_observability) to the granular
api/<module> entries. Add a consumer-facing README.md (tier decision table + recipes) and
refresh AGENTS.md (§1 facade, §2 generic tool seam, §5 layout, §6 replacement-readiness).

No runtime behavior change.
