+++ frontend/src/scenes/inbox/components/detail/AgentRunDetail.tsx
@@ -20,7 +20,8 @@ import { maxGlobalLogic } from 'scenes/max/maxGlobalLogic'
 import { SANDBOX_BIND_TASK_PARAM } from 'scenes/max/maxLogic'
 import { urls } from 'scenes/urls'
 
-import { isTerminalRunStatus, RunViewer } from 'products/posthog_ai/frontend'
+import { isTerminalRunStatus } from 'products/posthog_ai/frontend/api/logics'
+import { RunViewer } from 'products/posthog_ai/frontend/api/run'
 import { Task, TaskRunStatus } from 'products/posthog_ai/frontend/types/taskTypes'
 
 import { inboxReportDetailLogic } from '../../logics/inboxReportDetailLogic'
