+++ frontend/src/scenes/inbox/components/detail/ArtefactTaskRun.tsx
@@ -6,7 +6,8 @@ import { LemonTag } from '@posthog/lemon-ui'
 import api from 'lib/api'
 import { identifierToHuman } from 'lib/utils/strings'
 
-import { isTerminalRunStatus, RunViewer } from 'products/posthog_ai/frontend'
+import { isTerminalRunStatus } from 'products/posthog_ai/frontend/api/logics'
+import { RunViewer } from 'products/posthog_ai/frontend/api/run'
 import { Task, TaskRunStatus } from 'products/posthog_ai/frontend/types/taskTypes'
 
 import { isCustomAgentTaskRun, taskRunTypeLabel, TaskRunArtefactContent } from './artefactTypes'
