+++ frontend/src/scenes/max/components/MaxWebAnalyticsNudge.tsx
@@ -4,7 +4,7 @@ import { LemonButton } from '@posthog/lemon-ui'
 
 import { iconForType } from '~/layout/panel-layout/ProjectTree/defaultTree'
 
-import { MessageTemplate } from 'products/posthog_ai/frontend'
+import { MessageTemplate } from 'products/posthog_ai/frontend/api/primitives'
 
 import { maxWebAnalyticsNudgeLogic } from '../logics/maxWebAnalyticsNudgeLogic'
 import { maxLogic, ThreadMessage } from '../maxLogic'
