+++ frontend/src/scenes/max/maxTypes.ts
@@ -245,7 +245,7 @@ export function isAgentMode(mode: unknown): mode is AgentMode {
 
 // `ToolCallMessage` now lives with the relocated sandbox renderer. Re-exported here so
 // the frozen LangGraph path and any in-flight branches keep resolving it from `maxTypes`.
-export type { ToolCallMessage } from 'products/posthog_ai/frontend'
+export type { ToolCallMessage } from 'products/posthog_ai/frontend/api/types'
 
 /**
  * Flat context attachment sent to the sandbox agent runtime (`agent_runtime === 'sandbox'`).
