+++ frontend/src/scenes/max/messages/MultiQuestionForm.tsx
@@ -2,7 +2,7 @@ import { IconCheck, IconX } from '@posthog/icons'
 
 import { MultiQuestionForm as MultiQuestionFormType } from '~/queries/schema/schema-assistant-messages'
 
-import { MessageTemplate } from 'products/posthog_ai/frontend'
+import { MessageTemplate } from 'products/posthog_ai/frontend/api/primitives'
 
 interface MultiQuestionFormRecapProps {
     form: MultiQuestionFormType
