+++ frontend/src/scenes/max/messages/UIPayloadAnswer.tsx
@@ -31,7 +31,7 @@ import {
     issueQueryOptionsLogic,
 } from 'products/error_tracking/frontend/components/IssueQueryOptions/issueQueryOptionsLogic'
 import { ERROR_TRACKING_SCENE_LOGIC_KEY } from 'products/error_tracking/frontend/scenes/ErrorTrackingScene/errorTrackingSceneLogic'
-import { MessageTemplate } from 'products/posthog_ai/frontend'
+import { MessageTemplate } from 'products/posthog_ai/frontend/api/primitives'
 
 import { isDangerousOperationResponse, normalizeDangerousOperationResponse } from '../approvalOperationUtils'
 import { DangerousOperationApprovalCard } from '../DangerousOperationApprovalCard'
