+++ frontend/src/scenes/max/messages/adapters/ErrorTrackingWidget.tsx
@@ -1,4 +1,4 @@
-import { GenericMcpToolRenderer, DataToolRow, type ToolRendererProps } from 'products/posthog_ai/frontend'
+import { GenericMcpToolRenderer, DataToolRow, type ToolRendererProps } from 'products/posthog_ai/frontend/api/tools'
 
 import { ErrorTrackingFiltersWidget } from '../UIPayloadAnswer'
 import { extractErrorTrackingResponse } from './extractors'
