+++ frontend/src/scenes/max/messages/adapters/SearchSessionRecordingsWidget.tsx
@@ -1,4 +1,4 @@
-import { GenericMcpToolRenderer, DataToolRow, type ToolRendererProps } from 'products/posthog_ai/frontend'
+import { GenericMcpToolRenderer, DataToolRow, type ToolRendererProps } from 'products/posthog_ai/frontend/api/tools'
 
 import { RecordingsWidget } from '../UIPayloadAnswer'
 import { extractRecordingFilters } from './extractors'
