+++ frontend/src/scenes/max/messages/adapters/registerMaxToolRenderers.test.ts
@@ -3,9 +3,20 @@
 // Renderer itself is an opaque chunk).
 import './registerMaxToolRenderers'
 
-import { toolRegistry } from 'products/posthog_ai/frontend'
+import { registerToolRenderers, toolRegistry, type ToolRegistryEntry } from 'products/posthog_ai/frontend/api/tools'
 
 describe('registerMaxToolRenderers', () => {
+    it('registerToolRenderers bulk-registers every entry it is handed', () => {
+        const Renderer = (() => null) as unknown as ToolRegistryEntry['Renderer']
+        const icon = null as unknown as JSX.Element
+        registerToolRenderers([
+            { key: '__test_alpha__', displayName: 'Alpha', icon, Renderer },
+            { key: '__test_beta__', displayName: 'Beta', icon, Renderer },
+        ])
+        expect(toolRegistry.lookup('__test_alpha__')?.displayName).toEqual('Alpha')
+        expect(toolRegistry.lookup('__test_beta__')?.displayName).toEqual('Beta')
+    })
+
     it.each([
         ['insight-create', 'Insight'],
         ['insight-update', 'Insight'],
