+++ products/ai_observability/frontend/ConversationDisplay/TranscriptBubbleStream.tsx
@@ -5,7 +5,7 @@ import { IconChevronRight } from '@posthog/icons'
 
 import { LemonMarkdown } from 'lib/lemon-ui/LemonMarkdown'
 
-import { MessageTemplate } from 'products/posthog_ai/frontend'
+import { MessageTemplate } from 'products/posthog_ai/frontend/api/primitives'
 
 import { CompatMessage } from '../types'
 import {
